% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentalDesignSimulation.R
\name{FutureExperimentSimulation}
\alias{FutureExperimentSimulation}
\title{Test future experimental design using simulated data}
\usage{
FutureExperimentSimulation(
  N_proteins = 300,
  N_rep = 3,
  N_Control_Rep = NULL,
  Concentrations = c(0, 1, 3, 10, 30, 100, 300, 1000, 3000),
  IC50_Prediction = FALSE
)
}
\arguments{
\item{concentrations}{list of drug concentrations (in nM scale) in sim experiment}

\item{rep}{number of replicates for each drug concentration}

\item{template}{real protein level data representing different level of interactions}

\item{outlier_prob}{probability of sample outlier}
}
\description{
Test future experimental design using simulated data
}
