% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_turnover_ratio_helper.R
\name{parse_timepoint}
\alias{parse_timepoint}
\title{Parse timepoint strings to numeric hours}
\usage{
parse_timepoint(time_strings)
}
\arguments{
\item{time_strings}{Character vector of timepoint labels}
}
\value{
Numeric vector of hours
}
\description{
Parse timepoint strings to numeric hours
}
\examples{
MSstatsResponse:::parse_timepoint(c("0hr", "1hr", "12hrs", "168hrs"))
# Returns: 0 1 12 168

}
\keyword{internal}
