% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentalDesignSimulation.R
\name{plot_hit_rate_MSstatsResponse}
\alias{plot_hit_rate_MSstatsResponse}
\title{Plot hit rates by category}
\usage{
plot_hit_rate_MSstatsResponse(results, rep_count, concentration_count)
}
\arguments{
\item{results}{output of interaction test}

\item{rep_count}{number of replicates per concentration in simulation}

\item{concentration_count}{number of concentrations in simulation}
}
\description{
Plot hit rates by category
}
