% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Isotonic_Fit.R
\name{plot_isotonic}
\alias{plot_isotonic}
\title{Plot Isotonic Regression Model}
\usage{
plot_isotonic(
  fit,
  ratio = TRUE,
  show_ic50 = FALSE,
  drug_name = NULL,
  protein_name = NULL,
  x_lab = "dose (nM)",
  y_lab = "log2 Intensity",
  title = NULL,
  legend = FALSE,
  theme_style = "classic",
  original_label = TRUE
)
}
\arguments{
\item{fit}{A model object returned by fit_isotonic_regression().}

\item{show_ic50}{Logical. If TRUE, adds vertical line and annotation for IC50.}

\item{x_lab}{Label for x-axis.}

\item{y_lab}{Label for y-axis.}

\item{title}{Title for the plot.}

\item{legend}{Logical. Show legend if TRUE.}

\item{theme_style}{ggplot2 theme name to apply (default = "classic").}

\item{original_label}{Logical. If TRUE, replace x-axis tick labels with original dose labels.}
}
\value{
A ggplot object.
}
\description{
Plot Isotonic Regression Model
}
