% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_turnover_ratio_helper.R
\name{selectTopNPeptides}
\alias{selectTopNPeptides}
\title{Example peptide selector: Top N by total signal}
\usage{
selectTopNPeptides(df, n = 10)
}
\arguments{
\item{df}{Data frame with BaseSequence and Intensity columns (for one protein)}

\item{n}{Number of top peptides to select}
}
\value{
Filtered data frame
}
\description{
Example peptide selector: Top N by total signal
}
\examples{
# Use as peptide_selector
ratios <- calculateTurnoverRatios(
  feature_data = quant_data$FeatureLevelData,
  peptide_selector = function(df) selectTopNPeptides(df, n = 10)
)

}
