% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentalDesignSimulation.R
\name{simulate_chemo_proteinlevel_nonparametric}
\alias{simulate_chemo_proteinlevel_nonparametric}
\title{Simulate chemoproteomics data at the protein level - non-parametric approach}
\usage{
simulate_chemo_proteinlevel_nonparametric(
  N_proteins = 3000,
  TP = 0.333,
  TW = 0.333,
  TN = 0.333,
  concentrations = c(0, 1, 3, 10, 30, 100, 300, 1000, 3000),
  rep = 3,
  seed = 3,
  var_tech = 0.4,
  control_rep = NULL,
  template = list(strong_interaction = data.frame(dose = c(), LogIntensities = c()),
    weak_interaction = data.frame(dose = c(), LogIntensities = c()), no_interaction =
    data.frame(dose = c(), LogIntensities = c())),
  outlier_prob = 0.05
)
}
\arguments{
\item{N_proteins}{number of proteins in sim}

\item{TP}{percent of strong interacting proteins in sim}

\item{TW}{percent of weak interacting proteins in sim}

\item{TN}{percent of non interacting proteins in sim}

\item{concentrations}{list of drug concentrations in sim experiment}

\item{rep}{number of replicates for each drug concentration}

\item{seed}{simulation seed}

\item{template}{real protein level data representing different level of interactions}

\item{outlier_prob}{probability of sample outlier}

\item{var_teach}{combined technical and biological variation}
}
\description{
Simulate chemoproteomics data at the protein level - non-parametric approach
}
