% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{MSstatsPrepareDoseResponseFit}
\alias{MSstatsPrepareDoseResponseFit}
\title{Prepare data for dose-response fitting with isotonic regression}
\usage{
MSstatsPrepareDoseResponseFit(
  data,
  dose_column = "dose",
  drug_column = "drug",
  protein_column = "Protein",
  log_abundance_column = "LogIntensities",
  transform_nM_to_M = NULL
)
}
\arguments{
\item{data}{A data.frame (e.g. data$ProteinLevelData from MSstats)}

\item{dose_column}{Name of the column containing dose values (e.g., "dose")}

\item{drug_column}{Name of column containing treatment name (e.g. drug name )}

\item{protein_column}{Name of the column containing protein identifiers (e.g., "Protein")}

\item{log_abundance_column}{Name of the column with log-transformed abundance values (e.g., "LogIntensities")}
}
\value{
A standardized data.frame with columns: dose, response, protein
}
\description{
Prepare data for dose-response fitting with isotonic regression
}
