% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC50_predicition.R
\name{PredictIC50}
\alias{PredictIC50}
\title{Predict IC50 (dose where response is 0.5) for each protein and drug}
\usage{
PredictIC50(
  data,
  n_samples = 1000,
  alpha = 0.1,
  increasing = FALSE,
  transform_dose = TRUE,
  ratio_response = TRUE,
  bootstrap = TRUE
)
}
\arguments{
\item{data}{A data frame with columns: protein, drug, dose, response.}

\item{n_samples}{Number of bootstrap samples. Default = 1000.}

\item{alpha}{Confidence level. Default = 0.05.}

\item{increasing}{Logical. If TRUE, fit a non-decreasing trend.}

\item{transform_dose}{Logical. If TRUE, applies log10(dose + 1) transformation. Default is TRUE.}

\item{ratio_response}{Logical. If TRUE, use ratio response; else use log2 scale.}

\item{bootstrap}{Logical. If FALSE, skip confidence interval bootstrap estimation and only return IC50.}
}
\value{
A data frame with columns: protein, drug, IC50, lower CI, upper CI.
}
\description{
Predict IC50 (dose where response is 0.5) for each protein and drug
}
