% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC50_predicition.R
\name{predictIC50}
\alias{predictIC50}
\title{Predict IC50 (dose where response is 0.5 of DMSO)}
\usage{
predictIC50(fit, target_response = 0.5)
}
\arguments{
\item{fit}{An object of class "isotonic_model" from fitIsotonicRegression().}

\item{target_response}{The target response value for prediction (default = 0.5).}
}
\value{
Estimated dose value corresponding to target response (IC50)
}
\description{
Predict IC50 (dose where response is 0.5 of DMSO)
}
